ai-services/requirements.txt
# ProctorAI — Python AI Services
# Install with: pip install -r requirements.txt

# Web framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Computer Vision
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0

# MediaPipe (face mesh, pose, etc.)
mediapipe==0.10.9

# YOLOv8
ultralytics==8.1.27

# Audio transcription
openai-whisper==20231117
torch==2.2.0
torchaudio==2.2.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
webrtcvad==2.0.10

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
httpx==0.27.0
aiofiles==23.2.1
loguru==0.7.2

# For base64 handling
requests==2.31.0
